pyproject.toml
[tool.poetry]
name = "heimdall"
version = "0.1.0"
description = ""
authors = ["chalupaul <me@chalupaul.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
mangum = "^0.14.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
aws-cdk-lib = "^2.18.0"
uvicorn = "^0.17.6"
taskipy = "^1.10.1"
doit = "^0.36.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
cdksynth = "cdk synth"
cdkdeploy = "cdk deploy --require-approval never --all"
buildlayers = "doit -f cdk/common/build_layers.py"